requirements.txt
apache-airflow-providers-google

python-dotenv

numpy
pyarrow
pandas
sqlalchemy 
psycopg2-binary


pyspark
google-cloud-storage
